+++ ask-smapi-model/ask_smapi_model/v2/skill/simulations/simulation.py
@@ -23,59 +23,37 @@ from enum import Enum
 if typing.TYPE_CHECKING:
     from typing import Dict, List, Optional, Union, Any
     from datetime import datetime
-    from ask_smapi_model.v1.skill.simulations.session import Session as Session_700bf076
-    from ask_smapi_model.v1.skill.simulations.device import Device as Device_88be9466
-    from ask_smapi_model.v1.skill.simulations.input import Input as Input_7307d1de
-    from ask_smapi_model.v1.skill.simulations.simulation import Simulation as Simulation_618c97c6
+    from ask_smapi_model.v2.skill.simulations.simulation_type import SimulationType as SimulationType_d7c182c4
 
 
-class SimulationsApiRequest(object):
+class Simulation(object):
     """
+    Simulation settings for the current simulation request. 
 
-    :param input: 
-    :type input: (optional) ask_smapi_model.v1.skill.simulations.input.Input
-    :param device: 
-    :type device: (optional) ask_smapi_model.v1.skill.simulations.device.Device
-    :param session: 
-    :type session: (optional) ask_smapi_model.v1.skill.simulations.session.Session
-    :param simulation: 
-    :type simulation: (optional) ask_smapi_model.v1.skill.simulations.simulation.Simulation
+
+    :param object_type: 
+    :type object_type: (optional) ask_smapi_model.v2.skill.simulations.simulation_type.SimulationType
 
     """
     deserialized_types = {
-        'input': 'ask_smapi_model.v1.skill.simulations.input.Input',
-        'device': 'ask_smapi_model.v1.skill.simulations.device.Device',
-        'session': 'ask_smapi_model.v1.skill.simulations.session.Session',
-        'simulation': 'ask_smapi_model.v1.skill.simulations.simulation.Simulation'
+        'object_type': 'ask_smapi_model.v2.skill.simulations.simulation_type.SimulationType'
     }  # type: Dict
 
     attribute_map = {
-        'input': 'input',
-        'device': 'device',
-        'session': 'session',
-        'simulation': 'simulation'
+        'object_type': 'type'
     }  # type: Dict
     supports_multiple_types = False
 
-    def __init__(self, input=None, device=None, session=None, simulation=None):
-        # type: (Optional[Input_7307d1de], Optional[Device_88be9466], Optional[Session_700bf076], Optional[Simulation_618c97c6]) -> None
-        """
+    def __init__(self, object_type=None):
+        # type: (Optional[SimulationType_d7c182c4]) -> None
+        """Simulation settings for the current simulation request. 
 
-        :param input: 
-        :type input: (optional) ask_smapi_model.v1.skill.simulations.input.Input
-        :param device: 
-        :type device: (optional) ask_smapi_model.v1.skill.simulations.device.Device
-        :param session: 
-        :type session: (optional) ask_smapi_model.v1.skill.simulations.session.Session
-        :param simulation: 
-        :type simulation: (optional) ask_smapi_model.v1.skill.simulations.simulation.Simulation
+        :param object_type: 
+        :type object_type: (optional) ask_smapi_model.v2.skill.simulations.simulation_type.SimulationType
         """
         self.__discriminator_value = None  # type: str
 
-        self.input = input
-        self.device = device
-        self.session = session
-        self.simulation = simulation
+        self.object_type = object_type
 
     def to_dict(self):
         # type: () -> Dict[str, object]
@@ -120,7 +98,7 @@ class SimulationsApiRequest(object):
     def __eq__(self, other):
         # type: (object) -> bool
         """Returns true if both objects are equal"""
-        if not isinstance(other, SimulationsApiRequest):
+        if not isinstance(other, Simulation):
             return False
 
         return self.__dict__ == other.__dict__